refactor: 날짜 관련 함수 분리
